Rename drilldown file
