+++ packages/client/hmi-client/src/components/workflow/ops/calibrate-ciemss/mod.ts
@@ -1,6 +1,6 @@
 import { CalibrationOperationCiemss as operation } from './calibrate-operation';
 import node from './tera-calibrate-node-ciemss.vue';
-import drilldown from './tera-calibrate-ciemss.vue';
+import drilldown from './tera-calibrate-ciemss-drilldown.vue';
 
 const name = operation.name;
